+++ src/TMSTableToTMSFile2.py
@@ -8,7 +8,7 @@ conn = psycopg2.connect(host="durga.jrc.org", database="dbdopa", user="usrdopa",
 cur = conn.cursor()
 cur.execute("SELECT tx,ty,z FROM public.pilotspeciesdata WHERE speciesid='" + speciesID + "'")
 rows=cur.fetchall()
-records=numpy.transpose(numpy.array(rows))
+records=numpy.transpose(numpy.array(rows,dtype="int32"))
 records=numpy.reshape(records,len(records[0])*3) #convert the arraytx,arrayty,arrayz to a single array
 f = open(r"E:\cottaan\My Documents\ArcGIS\ID" + speciesID + ".npy",'wb')
 #f = open("/srv/www/htdocs/eSpecies/SpeciesRichness.npy",'wb')
